[Flat] Kernel refactor
