+++ src/ops/kernels/flat_kernels.cpp
@@ -13,14 +13,42 @@
  * limitations under the License.
  */
 
-#include "flexflow/ops/flat.h"
+#include "flexflow/ops/kernels/flat_kernels.h"
 #include "flexflow/utils/hip_helper.h"
 #include <hip/hip_runtime.h>
 
 namespace FlexFlow {
 
+namespace Kernels {
+namespace Flat {
+
 /*static*/
-void Flat::forward_kernel(float const *input_ptr,
+void forward_kernel_wrapper(float const *input_ptr,
+                                  float *output_ptr,
+                                  size_t num_elements) {
+  hipStream_t stream;
+  checkCUDA(get_legion_stream(&stream));
+  forward_kernel(input_ptr, output_ptr, num_elements, stream);
+  // checkCUDA(hipDeviceSynchronize());
+}
+
+void backward_kernel_wrapper(float *input_grad_ptr,
+                                   float const *output_grad_ptr,
+                                   size_t num_elements) {
+  hipStream_t stream;
+  checkCUDA(get_legion_stream(&stream));
+  backward_kernel(input_grad_ptr, output_grad_ptr, num_elements, stream);
+  // checkCUDA(hipMemcpyAsync(acc_input_grad.ptr, acc_output_grad.ptr,
+  //                           acc_input_grad.rect.volume() * sizeof(float),
+  //                           hipMemcpyDeviceToDevice));
+  // checkCUDA(hipDeviceSynchronize());
+}
+
+
+namespace Internal {
+
+/*static*/
+void forward_kernel(float const *input_ptr,
                           float *output_ptr,
                           size_t num_elements,
                           hipStream_t stream) {
@@ -31,17 +59,7 @@ void Flat::forward_kernel(float const *input_ptr,
                            stream));
 }
 
-/*static*/
-void Flat::forward_kernel_wrapper(float const *input_ptr,
-                                  float *output_ptr,
-                                  size_t num_elements) {
-  hipStream_t stream;
-  checkCUDA(get_legion_stream(&stream));
-  Flat::forward_kernel(input_ptr, output_ptr, num_elements, stream);
-  // checkCUDA(hipDeviceSynchronize());
-}
-
-void Flat::backward_kernel(float *input_grad_ptr,
+void backward_kernel(float *input_grad_ptr,
                            float const *output_grad_ptr,
                            size_t num_elements,
                            hipStream_t stream) {
@@ -57,16 +75,8 @@ void Flat::backward_kernel(float *input_grad_ptr,
                      alpha);
 }
 
-void Flat::backward_kernel_wrapper(float *input_grad_ptr,
-                                   float const *output_grad_ptr,
-                                   size_t num_elements) {
-  hipStream_t stream;
-  checkCUDA(get_legion_stream(&stream));
-  Flat::backward_kernel(input_grad_ptr, output_grad_ptr, num_elements, stream);
-  // checkCUDA(hipMemcpyAsync(acc_input_grad.ptr, acc_output_grad.ptr,
-  //                           acc_input_grad.rect.volume() * sizeof(float),
-  //                           hipMemcpyDeviceToDevice));
-  // checkCUDA(hipDeviceSynchronize());
-}
 
-}; // namespace FlexFlow
+} // namespace Internal
+} // namespace Flat
+} // namespace Kernels
+} // namespace FlexFlow
